Create Ping Pong
this program here is of ping_pong game. There is small error in this program, which you will automatically find out when you will run the code. I give this oppurtunity to you if  you are able to find out my error or not. If not rech me out and i will be happy to help you .